pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "readonly-db-mcp"
version = "0.1.0"
description = "Read-only MCP server for PostgreSQL and ClickHouse — safe database access for AI agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Passendo" },
]
keywords = ["mcp", "postgresql", "clickhouse", "read-only", "ai", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "mcp>=1.2.0",
    "sqlglot>=26.0",
    "asyncpg>=0.29",
    "clickhouse-connect>=0.7",
    "python-dotenv>=1.0",
]

[project.scripts]
readonly-db-mcp = "readonly_db_mcp.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.hatch.build.targets.wheel]
packages = ["src/readonly_db_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 120
